Add files via upload
During the process, I was able to successfully create a bot account. Also download Arduino, installed Arduino telegram bot library onto my computer and add to the Arduino library, and also Auduino ,initially the ESP32 cam was not connecting to the sketch, I realized I had picked the wrong version of the ESP32CAM on the tools, when I rectified that and finally connected.

My only issue was that telegram bot would not respond to the /start and / flash command.